+++ apps/web/src/components/overview-bottom-nav.tsx
@@ -27,10 +27,14 @@ export function OverviewBottomNav() {
   );
 }
 
-function NavButton({ icon: Icon, onClick = null, selected = false }) {
+export function NavButton({ icon: Icon, onClick = null, selected = false, variant = 'black' }) {
   const containerClassname = useClassName(
     'rounded-full p-2 flex items-center justify-center transition-colors delay-50 cursor-pointer',
-    selected ? 'bg-gray-900 text-white' : 'bg-transparent text-gray-900',
+    selected
+      ? 'bg-gray-900 text-white'
+      : variant === 'black'
+        ? 'bg-transparent text-gray-900'
+        : 'bg-transparent text-white',
   );
   return (
     <button